Create region_replicate_music.lua
Submitting on behalf of Michael McLennan.

Jake notes: This is a simple but elegant little script! I wish it actually worked like the Sibelius/Dorico version, where the repetition is at any level, and not just at the measure level. Maybe it wouldn't be hard to adapt to that... But for now, this is pretty slick!