Create hotkey_script_palettes.lua
So what the hey - this is all the work I’m going to do on this script and despite the redundant [Enter] keys I kinda like the way it works. Will post it in case it is useful for others or inspires some other approach. (And “hotkey” is a more informative title than “macro key” - thanks @aaron)