Create ostinato_maker.lua
Copy the current selection (of any length) and paste it consecutively to the right a nominated number of times.

I’m not sure this warrants inclusion in the repo since much the same thing can be done with “Paste Multiple”, though this is simpler and more intuitive to use. It’s also a natural successor to and considerable improvement upon “region_replicate_music” which is already included.