Update page_format_wizard.lua
Implemented a fix for the problems some people were having on Facebook.

I'm not THRILLED with this solution because while I understand how it fixes the problem, I don't understand why the problem was happening in the first place and would have liked to address that. But since one of those folks got it working I present it here...