Update pitch_entry_keep_delete.lua
Changed `Popups` to `RadioButtonGroups` for operational clarity, improved Markdown for website listing, and added __Script Info__ dialog through `show_notes_dialog()`. This PR is subject to @Jwink ’s approval since he generously allowed me to subvert his original script submission.

But the real reason for this PR is to dig deeper into Robert’s assertion that my __Key Commands__ are “hacks”. This is an ideal exemplar because there’s only one `CtrlEdit` (`number`) with a tightly prescribed format.

We read `CtrlEdit` contents at will, provide initial values, are told by the _Handler_ when they change, and may need subsequently to change their contents for any number of reasons. That’s all that my _Key Commands_ do. In this case the `number` control must only be numeric, and if the User accidentally types “z” we must either immediately message an error or cancel the activity (which could be confusing). So why not inhibit alphabet entry in the first place? When any new key is typed in the control just ignore it unless it’s numeric. Is that a hack? It’s certainly very handy for the user.

Similarly it should only be one digit (unless you deal regularly with 19-note chords). If the user types a second digit just use that one instead so they don’t need to hit the delete key first. This is exactly the same behaviour as when the `control` is highlighted. Which it probably is anyway. That feels a tad hack-ish, but is it really? Ctrl handle alert -> read value -> write revised value. Standard `CtrlEdit` handling.

So now we have all these left-over alphabetics … why not do something useful with them? Scripts can make all types of internal decisions about contents, enabling controls, re-drawing dialog contents etc, for any set of reasons. (Just like Robert’s incredibly clever `class browser` does).  If a key press is unwanted as __value__ input then surely it’s free to trigger anything else. Is that a hack?