Create music_spacing_hack.lua
This may not be deemed suitable for the REPO, There's a couple of Music Spacing options that I change frequently for different spacing scenarios. This is a little hack using a couple of hotkeys to do that very quickly, without a mouse and without navigating the whole Document → Document Options → Music Spacing → Avoid Collision of → menu/dialog system.