Update rest_slider.lua
I've pondered this at considerable length and this version of the script - __MODAL ONLY__ - is the way that it __ought to__ operate. _Modeless_ operation is fundamentally antithetical to the assumptions of the script's workflow method, and the complexity of coping with changing selection and layer number while the script remains active creates an unnecessary quagmire with little gain. The modal approach just makes a cleaner and more sensible operation, leaving Finale's inbuilt plugin clean-up to handle the undo element.